refactor: move routes into /routes (#13)
- refactor: routes and tests moved into their own module

- fix: put caching back in meetup route

- fix(eventbrite): allow dead_code in entire module

Module needs a lot of refactoring. Disabled warnings for now

- refactor: move tests into the same file as source code

- fix: remove cache module
