+++ backend/src/eventbrite/user.rs
@@ -1,3 +1,4 @@
+#![allow(dead_code)]
 use serde::{Deserialize, Serialize};
 use std::error::Error;
